[Networking] Remove certain networking classes from netbase
